Feat & Update :
Added a ft_printf and Gnl to my so_long.
Modified the printf so that it does not have to use libft.
Added the MLX library to the framework.
Fixed my makefile that didnt compile the mlx library.
Compiled the overall framework of so_long.
Modified the makefile so that it can compile both gnl and printf.